Renamed generic_mesh.hpp to mesh.hpp
