+++ liblava/resource.hpp
@@ -10,6 +10,5 @@
 #include <liblava/resource/buffer.hpp>
 #include <liblava/resource/format.hpp>
 #include <liblava/resource/image.hpp>
-// #include <liblava/resource/mesh.hpp>
-#include <liblava/resource/generic_mesh.hpp>
+#include <liblava/resource/mesh.hpp>
 #include <liblava/resource/texture.hpp>
