--- liblava/resource/mesh.cpp
@@ -1,12 +1,0 @@
-/**
- * @file         liblava/resource/mesh.cpp
- * @brief        Vulkan mesh
- * @authors      Lava Block OÜ and contributors
- * @copyright    Copyright (c) 2018-present, MIT License
- */
-
-#include <liblava/resource/mesh.hpp>
-
-namespace lava {
-
-} // namespace lava
